pyproject.toml
[project]
name = "monitoring-alertmanager"
version = "1.0"
requires-python = ">=3.12,<3.14" # T408108
dependencies = [
    "PyYaml",
]

[tool.poetry.group.test.dependencies]
tox = "*"
tox-current-env = "*"
black = "*"
ruff = "*"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[tool.poetry.build]
generate-setup-file = false
script = "scripts/install.py"
